fixed favorite by user issues

+++ Pokedex/ViewControllers/RandomViewController.swift
@@ -188,37 +188,37 @@ class RandomViewController: UIViewController {
     func removeFavorite() {
         
         // Create a background queue for performing the deletion operation
-            let queue = DispatchQueue.global(qos: .background)
+        let queue = DispatchQueue.global(qos: .background)
             
-            queue.async { [weak self] in
-                let query = PokemonFavoriteEntry.query()
-                    .include("user")
+        queue.async { [weak self] in
+            let query = PokemonFavoriteEntry.query()
+                .include("user")
                 
-                // Fetch objects defined in query (async)
-                query.find { [weak self] result in
-                    // Perform the deletion operation on the background queue
-                    switch result {
-                    case .success(let entries):
-                        for entry in entries {
-                            if(entry.pokemonID == self?.natDexNum) {
-                                do {
-                                    try entry.delete()
-                                    // remove pokemon from Favorites list
-                                    self?.delegate!.removedFavorite(item: entry)
-                                } catch {
+            // Fetch objects defined in query (async)
+            query.find { [weak self] result in
+                // Perform the deletion operation on the background queue
+                switch result {
+                case .success(let entries):
+                    for entry in entries {
+                        if(entry.pokemonID == self?.natDexNum && entry.user == User.current) {
+                            do {
+                                try entry.delete()
+                                // remove pokemon from Favorites list
+                                self?.delegate!.removedFavorite(item: entry)
+                            } catch {
                                     // Handle any errors that may occur during deletion
-                                    self?.showAlert(description: error.localizedDescription)
-                                }
-                                return
+                                self?.showAlert(description: error.localizedDescription)
                             }
+                            return
                         }
-                        
-                    case .failure(let error):
-                        // Handle any errors that may occur during query execution
-                        self?.showAlert(description: error.localizedDescription)
                     }
+                        
+                case .failure(let error):
+                    // Handle any errors that may occur during query execution
+                    self?.showAlert(description: error.localizedDescription)
                 }
             }
+        }
     }
     
     //checks if current Pokemon is favorited by user
@@ -226,14 +226,13 @@ class RandomViewController: UIViewController {
         
         let query = PokemonFavoriteEntry.query()
             .include("user")
-            .order([.descending("createdAt")])
 
         // Fetch objects defined in query (async)
         query.find { [weak self] result in
             switch result {
             case .success(let entries):
                 for entry in entries {
-                    if(entry.pokemonID == self?.natDexNum) {
+                    if(entry.pokemonID == self?.natDexNum && entry.user == User.current) {
                         self?.favoriteBtn.setImage(UIImage(systemName: "star.fill"), for: .normal)
                         return
                     }
